+++ src/main/scala/xitrum/handler/NetOption.scala
@@ -1,10 +1,11 @@
 package xitrum.handler
 
 import java.net.InetSocketAddress
-import scala.util.control.NonFatal
 
 import io.netty.bootstrap.ServerBootstrap
-import io.netty.channel.ChannelOption
+import io.netty.channel.{ChannelFuture, ChannelFutureListener, ChannelOption}
+import io.netty.channel.nio.NioEventLoopGroup
+
 import xitrum.Config
 
 object NetOption {
@@ -20,7 +21,7 @@ object NetOption {
   }
 
   /** Stops the JVM process if cannot bind to the port. */
-  def bind(service: String, bootstrap: ServerBootstrap, port: Int) {
+  def bind(service: String, bootstrap: ServerBootstrap, port: Int, bossGroup: NioEventLoopGroup, workerGroup: NioEventLoopGroup) {
     val ic = Config.xitrum.interface
 
     val addr = ic match {
@@ -28,22 +29,37 @@ object NetOption {
       case Some(hostnameOrIp) => new InetSocketAddress(hostnameOrIp, port)
     }
 
-    try {
-      bootstrap.bind(addr)
-    } catch {
-      case NonFatal(e) =>
-        val msg = ic match {
-          case None =>
-            ("Could not open port %d for %s server. " +
-            "Check to see if there's another process already running on that port.")
-            .format(port, service)
+    val portOpenFuture = bootstrap.bind(addr)
+    portOpenFuture.awaitUninterruptibly()  // Wait for the port to be opened
 
-          case Some(hostnameOrIp) =>
-            ("Could not open %s:%d for %s server. " +
-            "Check to see if there's another process already running at that address.")
-            .format(hostnameOrIp, port, service)
-        }
-        Config.exitOnStartupError(msg, e)
+    // Handle java.net.BindException
+    // https://github.com/netty/netty/issues/1864
+    // https://github.com/mauricio/postgresql-async/issues/68
+    if (!portOpenFuture.isSuccess) {
+      val msg = ic match {
+        case None =>
+          ("Could not open port %d for %s server. " +
+          "Check to see if there's another process already running on that port.")
+          .format(port, service)
+        case Some(hostnameOrIp) =>
+          ("Could not open %s:%d for %s server. " +
+          "Check to see if there's another process already running at that address.")
+          .format(hostnameOrIp, port, service)
+      }
+
+      bossGroup.shutdownGracefully()
+      workerGroup.shutdownGracefully()
+      Config.exitOnStartupError(msg, portOpenFuture.cause)
     }
+
+    // Connection established successfully
+
+    // Graceful shutdown when the socket is closed
+    portOpenFuture.channel.closeFuture.addListener(new ChannelFutureListener {
+      override def operationComplete(future: ChannelFuture) {
+        bossGroup.shutdownGracefully()
+        workerGroup.shutdownGracefully()
+      }
+    })
   }
 }